Reduce duplication some more, add countOfTokenInValue helper function to avoid repeated pattern
